+++ Elemental/Components/AeTable/CustomRowComparer.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Elemental.Components
+{
+    public class CustomRowComparer<T> : IComparer<T>
+    {
+        private Comparison<T> comparison;
+
+        public CustomRowComparer(Comparison<T> comparison)
+        {
+            if (comparison == null)
+                throw new ArgumentNullException(nameof(comparison));
+            this.comparison = comparison;
+        }
+
+        public int Compare(T x, T y)
+        {
+            return comparison(x, y);
+        }
+    }
+}
+
